build: generate vimfuncs at build time
This has a few advantages: we don't have to include the generated file in the
source tree, and using `include!()` avoids rustfmt and other tools checking the
file.

The location of the generated file can also be customized by setting `OUT_DIR`,
but cargo picks an appropriate temporary place in the target/ directory by
default.

Also set the default workspace members to the crates which have tests so I don't
have to forget using `--workspace` all the time.